Wait for device to connect to ADB
Round trip latency flow starts as soon as the device is in online state.
It is possible that the device might not have connected to the ADB yet
and the attempt to run "cat" for the latency will result in "device not
found" or "device offline" error.
Wait for the device to connect to ADB and be online to accept the ping

Bug: N/A
Test: Manual
Change-Id: I8d0bf90456b416b30817ebe5c854d8ba10781c52