Compute action enabled state eagerly
Using WhileSubscribed StateFlows means that presentation.value
will provide the initial value in the absence of a persistent
subscriber, rather than starting the flow and waiting for the first
value to arrive (even though that would be immediate since it's derived
from other StateFlows). Thus, we need to combine the flows eagerly
instead.

As a result, Wipe Data and Delete were always enabled from the UI, which
checked presentation.value in AnAction.update(). (We actually do want
these actions always enabled, but the UI needs to stop the devices first
if they are active.)

Bug: 286923523
Test: updated
Change-Id: Ifee8bb4a484df5e2374e20171eedccb8b9d0832f