CommunicationDeviceDetector: simplify, add comments
CommunicationDeviceDetector is used as a reference example of a detector
that uses LintMaps and checkPartialResults. This change makes some minor
improvements. We now use map.size and remove the unnecessary numSetCalls
field. The comments have been updated to use "incident" instead of
"issue", and we now explain how checkPartialResults may be called for
each Issue that the detector handles.

Bug: 260714563
Test: N/A
Change-Id: I1fffdeca6e390c4093525550a4fa5893dffc30a4