adblib: replace `flow` usages with `receive`
Using `JdwpPacketReceiver.flow()` is more convenient than
`JdwpPacketReceiver.receive()` but is less efficient, as
`JdwpPacket` payload need to be copied in memory
when using the `flow()` method.

This change replaces performance sensitive usages of
`flow()` with calls to `receive()`, while at the same time
introduces extension to `receive()` that make it more
convenient to use (e.g. `receiveFirst()`).

Bug: 188067646
Test: Included
Change-Id: Ibf5319f6cbca8222673ac041078b9ba6af44fb28