adblib: Add support for `JdwpSessionPipeline`
Add support for a `JdwpSessionPipeline` extension to the
debugging API. This will be useful when adding support for
the JDWP Speculative Cache (aka "scache") library which
allows speeding up JDWP traffic on low latency connections.

A `JdwpSessionPipeline` is a pipeline between JDWP
session that is proxied by `adblib` between an external
debugger (e.g. Android Studio) and an Android JDWP
process.

The pipeline always starts with a JDWP session pipe to/from
the external Debugger. Additional pipeline can be chained
together using the `JdwpSessionPipelineFactory`
extension point. The JDWP scache support will be
added trough this extension point.

In summary, here is how things are hooked up:

* External Debugger (e.g. Android Studio)
<=> (JDWP socket)
* DebuggerSessionPipeline
   <-> (pipeline)
   (other JdwpSessionPipeline #1, e.g. scache)
   <-> pipeline)
   (other JdwpSessionPipeline #2)
   <-> pipeline)
* adblib proxy
<=> (JDWP socket)
*Android Process on Android device

Bug: 188067646
Test: Existing
Change-Id: Idcc70c852beb54bbf6daf5a206511bc0c92d025f