fakeadb: Track active device commands
Add a way to track currently active commands for a given
device so that these commands can be cancelled (i.e.
socket is closed) when FakeAdb needs to simulate a
device disconnection (i.e. when FakeAdb.disconnectDevice)
is called.

This allows enabling an `adblib` test that relies on this
behavior.

Bug: 188067646
Test: Existing + included
Change-Id: I2dc0db12a491ac27c0af294416691aa58df1d0bc