Deduplicate issues around targetSdkVersion
We recently created manifest-specific issue id's for the
targetSdkVersion warnings in manifest files. This means that existing
baselines (and disable configurations) would no longer apply to manifest
files (unless updated, as was done in one of the tests), and there's a
different mechanism we can use to share issue id's between Gradle and
Manifest files which this CL uses.

Also switches the quickfix implementation for updating targetSdkVersion
in manifest files to the XML attribute setter implementation.

Test: Existing (updated) unit tests
Bug: None
Change-Id: Ic0f997a207987765bd105e64c943add752c6f622