adblib: Share `JdwpProcess` between `AdbSession`
One of the last incremental step to fixing b/292257159:
`JdwpProcess` instances are shared (in a map) in a
single `AdbSession`, but can also be shared from a
different `AdbSession` though the same
`SharedJdwpSessionProviderDelegateSessionFinder` extension
point defined in a previous change.

This change leaves a few unused classes, that will be
deleted in a follow up change (to make diff easier to read).

Bug: 292257159
Test: Existing
Change-Id: I5957c869ae38e7ea6d81d74d535101a0c69b7122